refactor(config): Create directory for configuration files
- Put config.py and middleware.py in a config directory under src
- Update main.py to point to correct paths for import
